+++ src/Routing/ResourceRegistrar.php
@@ -19,6 +19,7 @@
 namespace CloudCreativity\LaravelJsonApi\Routing;
 
 use Closure;
+use CloudCreativity\LaravelJsonApi\Api\Api;
 use CloudCreativity\LaravelJsonApi\Api\Repository;
 use CloudCreativity\LaravelJsonApi\Api\ResourceProviders;
 use Illuminate\Contracts\Routing\Registrar;
@@ -66,29 +67,32 @@ class ResourceRegistrar
      */
     public function api($apiName, array $options, Closure $routes)
     {
-        $options = $this->pushMiddleware($apiName, $options);
-        $api = $this->apiGroup($apiName, $options);
-        $providers = $this->apiProviders($apiName);
+        $api = $this->apiRepository->createApi($apiName);
+        $groupOptions = $this->groupOptions($api);
 
-        $this->router->group($options, function () use ($api, $routes, $providers) {
-            $routes($api, $this->router);
-            $providers->mountAll($api, $this->router);
+        $this->router->group($groupOptions, function () use ($api, $options, $routes) {
+            $group = new ApiGroup($this->router, $api, $options);
+
+            $this->router->group($options, function () use ($group, $routes) {
+                $routes($group, $this->router);
+            });
+
+            $providers = $this->apiProviders($api->getName());
+            $providers->mountAll($group, $this->router);
         });
     }
 
     /**
-     * @param $apiName
-     * @param array $options
+     * @param Api $api
      * @return array
      */
-    protected function pushMiddleware($apiName, array $options)
+    protected function groupOptions(Api $api)
     {
-        $middleware = (array) array_get($options, 'middleware');
-        array_unshift($middleware, "json-api:$apiName");
-
-        $options['middleware'] = $middleware;
-
-        return $options;
+        return [
+            'middleware' => 'json-api:' . $api->getName(),
+            'as' => $api->getUrl()->getName(),
+            'prefix' => $api->getUrl()->getNamespace(),
+        ];
     }
 
     /**